requirements.txt
aioredis==1.2.0
asgiref==3.1.2
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.6.2
Automat==0.7.0
backcall==0.1.0
cffi==1.12.3
channels==2.2.0
channels-redis==2.4.0
constantly==15.1.0
cryptography==2.7
daphne==2.3.0
decorator==4.4.2
dj-database-url==0.5.0
Django==2.2.2
django-heroku==0.3.1
djangorestframework==3.11.0
gunicorn==20.0.4
hiredis==1.0.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
ipdb==0.13.2
ipython==7.15.0
ipython-genutils==0.2.0
jedi==0.17.0
msgpack==0.6.1
mysqlclient==1.4.2.post1
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
pkg-resources==0.0.0
prompt-toolkit==3.0.5
psycopg2==2.8.5
psycopg2-binary==2.8.5
ptyprocess==0.6.0
pycparser==2.19
Pygments==2.6.1
PyHamcrest==1.9.0
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
traitlets==4.3.3
Twisted==19.2.1
txaio==18.8.1
wcwidth==0.2.3
whitenoise==5.1.0
zope.interface==4.6.0
